Light gauge PoC.  Working for the most part, but I'll be doing a few improvements for render layering (curse that vignette) and a few other things

+++ src/main/java/com/fantasticsource/dynamicstealth/server/senses/sight/Sight.java
@@ -463,7 +463,7 @@ public class Sight
         return -777;
     }
 
-    private static double lightLevelTotal(World world, Vec3d vec)
+    private static int lightLevelTotal(World world, Vec3d vec)
     {
         BlockPos blockpos = new BlockPos(vec);
         if (!world.isAreaLoaded(blockpos, 1)) return 0;
